REFSW: Split to _base and _sw, in  different files
